Remove unused classes and dependencies in feature:common module
