+++ app/src/main/java/akio/apps/myrun/MyRunApp.kt
@@ -2,13 +2,13 @@ package akio.apps.myrun
 
 import akio.apps.myrun.data.tracking.api.RouteTrackingState
 import akio.apps.myrun.data.tracking.api.model.RouteTrackingStatus
+import akio.apps.myrun.di.AppComponent
+import akio.apps.myrun.di.DaggerAppComponent
 import akio.apps.myrun.feature.base.AppNotificationChannel
 import akio.apps.myrun.feature.configurator.ConfiguratorFacade
 import akio.apps.myrun.feature.tracking.RouteTrackingService
 import akio.apps.myrun.log.CrashReportTree
 import akio.apps.myrun.log.MyDebugTree
-import akio.apps.myrun.di.AppComponent
-import akio.apps.myrun.di.DaggerAppComponent
 import akio.apps.myrun.worker.AppMigrationWorker
 import akio.apps.myrun.worker.UpdateUserRecentPlaceWorker
 import android.app.Application
