+++ app/src/main/java/akio/apps/myrun/feature/home/HomeScreen.kt
@@ -2,7 +2,7 @@ package akio.apps.myrun.feature.home
 
 import akio.apps.myrun.R
 import akio.apps.myrun.data.activity.api.model.BaseActivityModel
-import akio.apps.myrun.feature.base.ext.px2dp
+import akio.apps.myrun.feature.base.ktx.px2dp
 import akio.apps.myrun.feature.base.ui.AppDimensions.AppBarHeight
 import akio.apps.myrun.feature.base.ui.AppDimensions.FabSize
 import akio.apps.myrun.feature.base.ui.AppTheme
