+++ app/src/main/java/akio/apps/myrun/feature/splash/SplashActivity.kt
@@ -2,7 +2,7 @@ package akio.apps.myrun.feature.splash
 
 import akio.apps.myrun.data.authentication.api.model.SignInSuccessResult
 import akio.apps.myrun.feature.base.DialogDelegate
-import akio.apps.myrun.feature.base.lifecycle.collectEventRepeatOnStarted
+import akio.apps.myrun.feature.base.ktx.collectEventRepeatOnStarted
 import akio.apps.myrun.feature.base.viewmodel.lazyViewModelProvider
 import akio.apps.myrun.feature.main.MainActivity
 import akio.apps.myrun.feature.registration.SignInActivity
