+++ feature/activity/src/main/java/akio/apps/myrun/feature/activitydetail/ActivityRouteMapActivity.kt
@@ -2,8 +2,8 @@ package akio.apps.myrun.feature.activitydetail
 
 import akio.apps.myrun.feature.activity.R
 import akio.apps.myrun.feature.base.BitmapUtils.createDrawableBitmap
-import akio.apps.myrun.feature.base.ext.dp2px
-import akio.apps.myrun.feature.base.ext.getColorCompat
+import akio.apps.myrun.feature.base.ktx.dp2px
+import akio.apps.myrun.feature.base.ktx.getColorCompat
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
