+++ feature/common/src/main/java/akio/apps/myrun/feature/base/ktx/ContextKtx.kt
@@ -1,0 +1,9 @@
+package akio.apps.myrun.feature.base.ktx
+
+import android.content.Context
+import androidx.annotation.ColorRes
+import androidx.core.content.ContextCompat
+
+fun Context.getColorCompat(@ColorRes color: Int): Int {
+    return ContextCompat.getColor(this, color)
+}
