--- feature/common/src/main/java/akio/apps/myrun/feature/base/permissions/PermissionUtils.kt
@@ -1,14 +1,0 @@
-package akio.apps.myrun.feature.base.permissions
-
-import android.content.Context
-import android.content.pm.PackageManager
-import androidx.core.content.ContextCompat
-
-object PermissionUtils {
-    fun arePermissionsGranted(context: Context, permissions: Array<String>): Boolean {
-        return permissions.all { permission ->
-            ContextCompat.checkSelfPermission(context, permission) ==
-                PackageManager.PERMISSION_GRANTED
-        }
-    }
-}
