+++ feature/profile/src/main/java/akio/apps/myrun/feature/profile/LinkStravaActivity.kt
@@ -2,8 +2,8 @@ package akio.apps.myrun.feature.profile
 
 import akio.apps.myrun.domain.launchcatching.Event
 import akio.apps.myrun.feature.base.DialogDelegate
-import akio.apps.myrun.feature.base.lifecycle.collectEventRepeatOnStarted
-import akio.apps.myrun.feature.base.lifecycle.collectRepeatOnStarted
+import akio.apps.myrun.feature.base.ktx.collectEventRepeatOnStarted
+import akio.apps.myrun.feature.base.ktx.collectRepeatOnStarted
 import akio.apps.myrun.feature.base.viewmodel.lazyViewModelProvider
 import akio.apps.myrun.feature.profile.wiring.DaggerLinkStravaComponent
 import android.content.Intent
