+++ feature/profile/src/main/java/akio/apps/myrun/feature/profile/ui/CropImageView.kt
@@ -1,7 +1,7 @@
 package akio.apps.myrun.feature.profile.ui
 
-import akio.apps.myrun.feature.base.ext.dp2px
-import akio.apps.myrun.feature.base.ext.toRect
+import akio.apps.myrun.feature.base.ktx.dp2px
+import akio.apps.myrun.feature.base.ktx.toRect
 import android.annotation.SuppressLint
 import android.content.Context
 import android.graphics.Bitmap
