+++ feature/registration/src/main/java/akio/apps/myrun/feature/registration/SignInActivity.kt
@@ -2,8 +2,8 @@ package akio.apps.myrun.feature.registration
 
 import akio.apps.myrun.data.authentication.api.model.SignInSuccessResult
 import akio.apps.myrun.feature.base.DialogDelegate
-import akio.apps.myrun.feature.base.lifecycle.collectEventRepeatOnStarted
-import akio.apps.myrun.feature.base.lifecycle.collectRepeatOnStarted
+import akio.apps.myrun.feature.base.ktx.collectEventRepeatOnStarted
+import akio.apps.myrun.feature.base.ktx.collectRepeatOnStarted
 import akio.apps.myrun.feature.base.viewmodel.lazyViewModelProvider
 import android.app.Activity
 import android.content.Context
