+++ feature/route/src/main/java/akio/apps/myrun/feature/route/ui/RoutePlanningActivity.kt
@@ -4,11 +4,11 @@ import akio.apps.myrun.data.location.api.model.LatLng
 import akio.apps.myrun.data.route.DaggerRouteDataComponent
 import akio.apps.myrun.feature.base.BitmapUtils
 import akio.apps.myrun.feature.base.DialogDelegate
-import akio.apps.myrun.feature.base.ext.dp2px
-import akio.apps.myrun.feature.base.ext.extra
-import akio.apps.myrun.feature.base.ext.getColorCompat
-import akio.apps.myrun.feature.base.lifecycle.collectEventRepeatOnStarted
-import akio.apps.myrun.feature.base.lifecycle.collectRepeatOnStarted
+import akio.apps.myrun.feature.base.ktx.collectEventRepeatOnStarted
+import akio.apps.myrun.feature.base.ktx.collectRepeatOnStarted
+import akio.apps.myrun.feature.base.ktx.dp2px
+import akio.apps.myrun.feature.base.ktx.extra
+import akio.apps.myrun.feature.base.ktx.getColorCompat
 import akio.apps.myrun.feature.base.viewmodel.lazyViewModelProvider
 import akio.apps.myrun.feature.route.R
 import akio.apps.myrun.feature.route.RoutePlanningViewModel
