+++ feature/tracking/src/main/java/akio/apps/myrun/feature/tracking/RouteTrackingViewModel.kt
@@ -15,8 +15,8 @@ import akio.apps.myrun.data.tracking.api.model.RouteTrackingStatus
 import akio.apps.myrun.domain.launchcatching.LaunchCatchingDelegate
 import akio.apps.myrun.domain.tracking.ClearRouteTrackingStateUsecase
 import akio.apps.myrun.domain.tracking.StoreTrackingActivityDataUsecase
+import akio.apps.myrun.feature.base.flowTimer
 import akio.apps.myrun.feature.tracking.model.RouteTrackingStats
-import akio.apps.myrun.log.flowTimer
 import akio.apps.myrun.worker.UploadStravaFileWorker
 import android.app.Application
 import android.graphics.Bitmap
